+++ src/ww/types.py
@@ -7,14 +7,14 @@ try:
 except NameError:
     unicode = str
 
+from typing import (Union, Iterable, Callable, Any, TypeVar, Hashable,  # noqa
+                    Generic, Iterator)
 
-from typing import (Union, Iterable, Callable, Any, TypeVar, Hashable,
-                    Generic, Iterator)  # noqa
-T = TypeVar('T')
-T2 = TypeVar('T2')
-C = TypeVar('C', bound=Callable)
-I = TypeVar('I', bound=Iterable)
-istr = Iterable[past_unicode]
-str_istr = Union[past_unicode, istr]
-str_or_callable = Union[past_unicode, Callable]
-str_istr_icallable = Union[past_unicode, Iterable[str_or_callable]]
+T = TypeVar('T')  # noqa
+T2 = TypeVar('T2')  # noqa
+C = TypeVar('C', bound=Callable)  # noqa
+I = TypeVar('I', bound=Iterable)  # noqa
+istr = Iterable[past_unicode]  # noqa
+str_istr = Union[past_unicode, istr]  # noqa
+str_or_callable = Union[past_unicode, Callable]  # noqa
+str_istr_icallable = Union[past_unicode, Iterable[str_or_callable]]  # noqa